rename debugger class
